Precalculating bitmaps. Slightly faster but not by much. Hopefully compatible with later animation sprites.

+++ app/src/main/java/com/example/pocketrocket/managers/ECSManager.kt
@@ -31,7 +31,7 @@ abstract class ECSManager(protected val callbackGameManager: GameManagerCallback
 
     fun resize(width: Int, height: Int) {
         val screenProperties = getScreenProperties()
-        if (width == screenProperties.width.toInt() && height == screenProperties.height.toInt()) return
+        if (width == screenProperties.width && height == screenProperties.height) return
         // Do stuff with stuff that cares about the screen dimensions...
     }
